# swagger/vega/api/v1/core_state_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "vega/api/v1/core_state_service.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/accounts": {
      "get": {
        "operationId": "ListAccounts",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListAccountsResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "party",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "market",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "CoreStateService"
        ]
      }
    },
    "/assets": {
      "get": {
        "operationId": "ListAssets",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListAssetsResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "asset",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "CoreStateService"
        ]
      }
    },
    "/delegations": {
      "get": {
        "operationId": "ListDelegations",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListDelegationsResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "party",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "node",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "epoch_seq",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "CoreStateService"
        ]
      }
    },
    "/markets": {
      "get": {
        "operationId": "ListMarkets",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListMarketsResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "market",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "CoreStateService"
        ]
      }
    },
    "/markets/data": {
      "get": {
        "operationId": "ListMarketsData",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListMarketsDataResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "market",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "CoreStateService"
        ]
      }
    },
    "/network/parameters": {
      "get": {
        "operationId": "ListNetworkParameters",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListNetworkParametersResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "network_parameter_key",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "CoreStateService"
        ]
      }
    },
...
    "/parties/stake": {
      "get": {
        "operationId": "ListPartiesStake",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListPartiesStakeResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "party",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "CoreStateService"
        ]
      }
    },
    "/proposals": {
      "get": {
        "operationId": "ListProposals",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListProposalsResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "proposal",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "proposer",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "CoreStateService"
        ]
      }
    },
...
    "/votes": {
      "get": {
        "operationId": "ListVotes",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListVotesResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "proposal",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "party",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "CoreStateService"
        ]
      }
    }
  },
  "definitions": {
...
    "apiv1Account": {
      "type": "object",
      "properties": {
        "party": {
          "type": "string"
        },
        "market": {
          "type": "string"
        },
        "balance": {
          "type": "string"
        },
        "asset": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    },
...
    "v1ListAccountsResponse": {
      "type": "object",
      "properties": {
        "accounts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiv1Account"
          }
        }
      }
    },
    "v1ListAssetsResponse": {
      "type": "object",
      "properties": {
        "assets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/vegaAsset"
          }
        }
      }
    },
    "v1ListDelegationsResponse": {
      "type": "object",
      "properties": {
        "delegations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/vegaDelegation"
          }
        }
      }
    },
    "v1ListMarketsDataResponse": {
      "type": "object",
      "properties": {
        "markets_data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/vegaMarketData"
          }
        }
      }
    },
    "v1ListMarketsResponse": {
      "type": "object",
      "properties": {
        "markets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/vegaMarket"
          }
        }
      }
    },
    "v1ListNetworkParametersResponse": {
      "type": "object",
      "properties": {
        "network_parameters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/vegaNetworkParameter"
          }
        }
      }
    },
...
    "v1ListPartiesStakeResponse": {
      "type": "object",
      "properties": {
        "parties_stake": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1PartyStake"
          }
        }
      }
    },
    "v1ListProposalsResponse": {
      "type": "object",
      "properties": {
        "proposals": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/vegaProposal"
          }
        }
      }
    },
...
    "v1ListVotesResponse": {
      "type": "object",
      "properties": {
        "votes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/vegaVote"
          }
        }
      }
    },
...
    "v1PartyStake": {
      "type": "object",
      "properties": {
        "party": {
          "type": "string"
        },
        "current_stake_available": {
          "type": "string"
        },
        "stake_linkings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1StakeLinking"
          }
        }
      }
    },
...
    "vegaAsset": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "Internal identifier of the asset"
        },
        "details": {
          "$ref": "#/definitions/vegaAssetDetails",
          "title": "The definition of the external source for this asset"
        }
      },
      "title": "The Vega representation of an external asset"
    },
...
    "vegaDelegation": {
      "type": "object",
      "properties": {
        "party": {
          "type": "string",
          "title": "Party which is delegating"
        },
        "node_id": {
          "type": "string",
          "title": "Node ID"
        },
        "amount": {
          "type": "string",
          "title": "Amount delegated"
        },
        "epoch_seq": {
          "type": "string",
          "title": "Epoch of delegation"
        }
      }
    },
...
    "vegaMarket": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "Unique identifier"
        },
        "tradable_instrument": {
          "$ref": "#/definitions/vegaTradableInstrument",
          "title": "Tradable instrument configuration"
        },
        "decimal_places": {
          "type": "string",
          "format": "uint64",
          "title": "Number of decimal places that a price must be shifted by in order to get a\ncorrect price denominated in the currency of the market, for example:\n`realPrice = price / 10^decimalPlaces`"
        },
        "fees": {
          "$ref": "#/definitions/vegaFees",
          "title": "Fees configuration"
        },
        "opening_auction": {
          "$ref": "#/definitions/vegaAuctionDuration",
          "title": "Auction duration specifies how long the opening auction will run (minimum\nduration and optionally a minimum traded volume)"
        },
        "continuous": {
          "$ref": "#/definitions/vegaContinuousTrading",
          "title": "Continuous"
        },
        "discrete": {
          "$ref": "#/definitions/vegaDiscreteTrading",
          "title": "Discrete"
        },
        "price_monitoring_settings": {
          "$ref": "#/definitions/vegaPriceMonitoringSettings",
          "title": "PriceMonitoringSettings for the market"
        },
        "liquidity_monitoring_parameters": {
          "$ref": "#/definitions/vegaLiquidityMonitoringParameters",
          "title": "LiquidityMonitoringParameters for the market"
        },
        "trading_mode": {
          "$ref": "#/definitions/MarketTradingMode",
          "title": "Current mode of execution of the market"
        },
        "state": {
          "$ref": "#/definitions/vegaMarketState",
          "title": "Current state of the market"
        },
        "market_timestamps": {
          "$ref": "#/definitions/vegaMarketTimestamps",
          "title": "Timestamps for when the market stay changes"
        }
      },
      "title": "Market definition"
    },
...
    "vegaNetworkParameter": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string",
          "title": "The unique key"
        },
        "value": {
          "type": "string",
          "title": "The value for the network parameter"
        }
      },
      "title": "Represents a network parameter on Vega"
    },
...
    "vegaProposal": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "Unique proposal identifier"
        },
        "reference": {
          "type": "string",
          "title": "Proposal reference"
        },
        "party_id": {
          "type": "string",
          "title": "Party identifier of the author (the party submitting the proposal)"
        },
        "state": {
          "$ref": "#/definitions/vegaProposalState",
          "title": "Proposal state - See (Proposal.State)[#vega.Proposal.State] definition"
        },
        "timestamp": {
          "type": "string",
          "format": "int64",
          "title": "Proposal timestamp for date and time (in nanoseconds) when proposal was submitted to the network"
        },
        "terms": {
          "$ref": "#/definitions/vegaProposalTerms",
          "title": "Proposal configuration and the actual change that is meant to be executed when proposal is enacted"
        },
        "reason": {
          "$ref": "#/definitions/vegaProposalError",
          "title": "A reason for the current state of the proposal, this may be set in case of REJECTED and FAILED statuses"
        },
        "error_details": {
          "type": "string",
          "description": "The detailed error associated to the reason."
        }
      },
      "title": "Governance proposal"
    },
...
    "vegaTradableInstrument": {
      "type": "object",
      "properties": {
        "instrument": {
          "$ref": "#/definitions/vegaInstrument",
          "title": "Instrument details"
        },
        "margin_calculator": {
          "$ref": "#/definitions/vegaMarginCalculator",
          "title": "Margin calculator for the instrument"
        },
        "log_normal_risk_model": {
          "$ref": "#/definitions/vegaLogNormalRiskModel",
          "title": "Log normal"
        },
        "simple_risk_model": {
          "$ref": "#/definitions/vegaSimpleRiskModel",
          "title": "Simple"
        }
      },
      "title": "Tradable Instrument definition"
    },
...
  }
}